Add support for tag inheritance.
This commit adds a new Plugin for handling "#+FILETAGS" lines that
attaches any tags found therein to the root element of OrgDataStructure.

It also adds a method `get_all_tags` to class OrgElement.Element
that will return the tags not only attached directly to it, but all
tags it inherits (either from parent nodes or from in-buffer settings).
Parameters to this method are provided that mimic the effects of the
Emacs org-mode variables `org-use-tag-inheritance` and
`org-tags-exclude-from-inheritance`.